requirements.txt
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.1.2
colorama==0.4.4
dominate==2.6.0
Flask-Bootstrap==3.3.7.1
Flask-WTF==1.0.1
Flask==2.1.1
idna==3.3
importlib-metadata==4.11.3
itsdangerous==2.1.2
Jinja2==3.1.1
MarkupSafe==2.1.1
pip==21.1.2
python-dotenv==0.20.0
requests==2.27.1
setuptools==57.0.0
urllib3==1.26.9
visitor==0.1.3
Werkzeug==2.1.1
wheel==0.36.2
WTForms==3.0.1
zipp==3.8.0
